fix: STT precision float32 → float16 (33x speedup)
float32 causes 108s transcription vs 3.2s with float16 on RTX 5070 Ti.
No quality difference for faster-whisper large-v3.

+++ api_v2.py
@@ -229,7 +229,7 @@ class STTRequest(BaseModel):
     output_folder: str
     model_path: str = "tools/asr/models/faster-whisper-large-v3"
     language: str = "auto"
-    precision: str = "float32"
+    precision: str = "float16"
 
 
 class DatasetFormattingRequest(BaseModel):
